requirements/production.txt
requests==2.32.3
playwright==1.48.0
beautifulsoup4==4.12.3
pyyaml==6.0.2
python-dotenv==1.0.1
psycopg2-binary==2.9.10
loguru==0.7.2